Move Twig functions and filters til actual Twig extension.
